pandu_supreme_v4 (1)/requirements.txt
fastapi
uvicorn[standard]
jinja2
httpx
websockets
python-dotenv
pandas
numpy
ta
pytelegrambotapi
aiofiles
beautifulsoup4
requests
